Added "daisy Chained" class heiarchy begun preliminary console outpup
"Daisy chaining"
Every object is given a reference to the object that contains it.

Organisms exist in an ecosystem and have a reference to that ecosystem
object.

The ecosystem exists in a Grid object and has a reference to that grid
object.

An animal can acsess the grid by acsessing the ecosystem and the
ecosystem acsessing the grid.